Fix downloaded files having trailing underscore
The file extension always had an ending double quote from incomplete
parsing, e.g. `.html"`. Any such characters forces (some?) browsers (on
some platforms?) to replace them with underscores, so you have a
downloaded file like `report.html_` (or `report.html"` if not replaced,
which is not helpful either). Strip (trim) it out and be done with it so
we can finally click to open.

P.S: This part of the code is not covered by automated (integration /
end-to-end) tests yet.