Upgrade to Python 3.8, but w/ Connexion 2
It appears there are two upgrade paths for us: (i) with connexion >=3
and (ii) with connexion <3. Connexion 3+ broke API, and applying some
basic refactoring according to their docs still gives us a 403 on the
API (meaning the routes couldn't be set up for some reason).

So, that leaves checking out what could be upgraded for connexion <3, as
even in that case there is significant dependency hell. It appears
there's a nice little spot that is indeed compatible with Python 3.8,
and that's with the few packages locked in this commit.