Soft-lock dependency versions
Lock to the latest _compatible_ packages of the lowest-supported Python
(3.8). Connexion 3+ and `pytest-bdd` 5+ both have breaking API changes
so they are not the actual "latest compatible". Because of Connexion,
Flask, along with `jsonschema` and `pyrsistent`, is kept at a
lower-than-actually-compatible version.

P.S: A compatible version in SemVer and PEP 440 usually means the patch
version is allowed to changed, but projects may not follow the same
versioning conventions (for e.g., if someone releases a breaking change
on a patch release). However, we rely on the wider Python community to
continue maintaining best practices.