Add support for injecting additional metadata
The current views (and reports) lack important metadata such as the file
names. Furthermore, the short-ID-based diff name only confuses users, so
only the original user-supplied name should be visible in user-facing
views. We therefore now inject more than just a description.

The internal RDF repository ID, which we can call the Dataset ID (to
distinguish it from the actual UID of the diff object), can accompany
the original name as secondary information. Additionally, we should
place old and new in that order for the user's intuition of a diff.

While we're at it, clean up the views a bit to make things look just a
tad little nicer (a big heading with the field name definitely looks
odd, and we certainly don't need to know at all times what the namespace
is or what the old and new version IDs are).