Fix deletions reporting multiple unexpected values
When a deleted value is part of a set of values, the report shows ALL of
the values as deleted, instead of just the one that was deleted. This is
a problem with the queries, as the deletions delta graph correctly shows
the deleted values only.

It turns out that in the deletion BGPs the values are always unbound
because of the anonymous variable `[]`. This brings in all values when
the final BGP for returning optional labels matches on all of them,
because of this loose binding.

In order to avoid this, simply use the named variable that is already
used in the last BGP. Doing this should have no side-effects. Please see
`diff-query-generator` for the relevant changes behind this.