Fix domain and range not appearing for OWL report
Ever since adding support for the OWL Core profile, in particular
semantic properties like `rdfs:domain` and `rdfs:range`, it has gone
unnoticed that changes related to these properties do not come up. This
was because in the corresponding queries there was a filter to check if
the language is the same for both the old and new data.

A more intricate query has been put in place in order to cater for when
there is no scope for such language tags to exist. However, it is not
known at this time why this is required in the first place. It needs to
be checked what is in fact the effect of having a different language
tag, or no language tag to a language tag, and vice-versa.