fix: freezing queries in some cases due to unbound variables
The cross-instance movement delta queries do not finish for certain
diffs (of a certain size), like ePO SHACL 4.x and 5.x. This is because
in commit e9d32e7863efc9ee02d54c264c1eac3ff46e81d4 "Update new profiles
to use fallback label properties" a blanket update was made for the
alternative label fallback OPTIONAL queries, which missed that in some
cases there were subqueries bound to `?newInstance` and `?oldInstance`
but `?instance` was used. This led to a cardinality explosion.

Fixed as per as per meaningfy-ws/diff-query-generator#32.