fix: creating empty dir in reports throws off the API
For a recent test it was chosen to create and write to a folder under
`rerports`. This throws off certain assumptions made by existing code,
testing or otherwise, and will fail invokations of the API in certain
cases, like when getting a diff. The code will try to look for a
`meta.json` which will fail in such cases. Revert to using a custom
folder for the purpose of the test and add a disclaimer about this in
the README.