chore(data): clean up sample vocabularies, add OWL+SHACL artefact
- Remove `*-no-import` files from `evaluation/` as the original problems
  should have disappeared by now.

- Add a combined OWL and SHACL artefact containing embedded SHACL data,
  for use cases that require the union of the two graphs in order to
fetch constraint information that would otherwise not be available in a
basic OWL ontology without restrictions (and because we do not support
advanced OWL constructs due to the challenge with arbitrary blank
nodes). This is currently mainly for domain, range and min/max
cardinality. While this was not part of the original `evaluation` it
made sense to keep it there, especially because it can be used for
profiling.

- Add a similar combined artefact to the test data, updating the
  existing ePO 4 sample.