feat(test): initial feature test for OWL diffing
Use the ePO 4.1 vs. 4.2 minimal sample and create a BDD test for the OWL
AP using Gherkin scenario outlines, with each example representing an
expected change. A pre-generated report is saved for use in this test if
not choosing to generate a report (this is the default for now).

The test currently takes the environment variables `BASE_URL` and
`REUSE_SAVED_REPORT` (valid strings 1 or 0, true or false). The former
allows overriding the default `localhost:4030` (e.g. if using Traefik it
is `api.localhost`), while the latter if true will run a diff once
(there is a caching mechanism to not repeat for all outlines).

The logic is somewhat rudimentary for now, as we hardcode certain
things, not least the prefixes, because the JSON report does not have
them, yet we would like to keep the Gherkin feature readable. The
algorithm for checking report items is also somewhat hardcoded, in the
sense that more change types would require more else blocks.

In this first attempt at Application Profile feature testing we start
out with a minimal set of cases and change types. Also to be noted, that
we use the CLI script to run the test, as it provides easy access to the
required sequence of operations (so we get a CLI test for free).